+++ Sources/SwiftMaskText/Info.plist
