+++ Sources/SwiftMaskText/Source/StringExtensions.swift
