--- Sources/SwiftMaskText/SwiftMaskText.swift
@@ -1,6 +1,0 @@
-public struct SwiftMaskText {
-    public private(set) var text = "Hello, World!"
-
-    public init() {
-    }
-}
