+++ Tests/SwiftMaskTextTests/Info.plist
