+++ Tests/SwiftMaskTextTests/SwiftMaskTextTests.swift
@@ -1,11 +1,36 @@
+//
+//  SwiftMaskTextTests.swift
+//  SwiftMaskTextTests
+//
+//  Created by Andre Santana Ferreira on 07/06/17.
+//  Copyright © 2017 com.moraisandre. All rights reserved.
+//
+
 import XCTest
 @testable import SwiftMaskText
 
-final class SwiftMaskTextTests: XCTestCase {
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-        XCTAssertEqual(SwiftMaskText().text, "Hello, World!")
+class SwiftMaskTextTests: XCTestCase {
+    
+    override func setUp() {
+        super.setUp()
+        // Put setup code here. This method is called before the invocation of each test method in the class.
     }
+    
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+        super.tearDown()
+    }
+    
+    func testExample() {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+    
+    func testPerformanceExample() {
+        // This is an example of a performance test case.
+        self.measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+    
 }
